fix: add lang in localStorage, add class
