+++ .eslintrc.json
@@ -14,6 +14,7 @@
     "import/extensions": ["error", "always"],
     "no-plusplus": ["error", { "allowForLoopAfterthoughts": true }],
     "no-param-reassign": ["error", { "props": false }],
-    "no-lonely-if": 0
+    "no-lonely-if": 0,
+    "class-methods-use-this": "off"
   }
 }
