+++ package.json
@@ -5,7 +5,7 @@
   "main": ".src/main.js",
   "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "eslint": "npx eslint ./src"
   },
   "repository": {
     "type": "git",
